refactor badges lambda function
